+++ Assets/External/Pathfinding/AStar/ManhattanHeuristic.cs
@@ -3,7 +3,7 @@ using Convenience.Geometry;
 
 namespace Pathfinding.AStar {
     public class ManhattanHeuristic : Heuristic {
-        public double Calculate(Point2D position, Point2D destination) {
+        public float Calculate(Point2D position, Point2D destination) {
             return Math.Abs(position.X - destination.X) + Math.Abs(position.Y - destination.Y);
         }
     }
